Add ensurePayed function
