+++ src/global.scss
@@ -1,0 +1,1 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700;900&display=swap');
