+++ src/variables.scss
@@ -1,0 +1,5 @@
+$font-family: 'Roboto', sans-serif;
+$border: 1px solid #babac0;
+$border-radius: 5px;
+$accent-color: #2a9cff;
+$background-color: #1c1c24;
